Added functionality for generating report items under each area. Still need to utilize subtotals method in the Fund controller to display subtotal information by area in the report view.

+++ FundEntities/Entities/Fund.cs
@@ -28,11 +28,11 @@ namespace FundEntities
 
         public string ResponsiblePerson { get; set; }
 
-        public decimal CurrentBudget { get; set; }
+        public int CurrentBudget { get; set; }
 
-        public decimal ProjectedExpenditures { get; set; }
+        public int ProjectedExpenditures { get; set; }
 
-        public decimal BudgetAdjustment { get; set; }
+        public int BudgetAdjustment { get; set; }
 
         public string BudgetAdjustmentNote { get; set; }
